Added reference OpenCL C++ bindings
OpenCL runtimes and SDKs will alwas ship with the necessaryy C API and
headers.  The C++ API, however, is optional. The NVIDIA CUDA Toolkit ini
particular does not ship with the C++ bindings.  Since the C++ API is only a
header-only wrapper arount the C API, then we can safely use the reference
implementation as a fallback of the vendor does not supply one.